Create more comprehensive AWS reference, clarify KRB_MANAGE_KRB5_CONF
The updates to the AWS reference conf file include the following
changes:

1. All instance types now inherit from a common-instance-properties
   block
2. Clarifying comments have been added in different sections
3. RDS is now used as the default database configuration and the
   corresponding code blocks for it have been uncommented
4. Kerberos configurations have been included as commented code blocks
5. CM Repository and parcel repository links are now uncommented
6. Additional services and roles have been added

Our comments surrounding KRB_MANAGE_KRB5_CONF in the AWS reference conf
file and the Azure Kerberos conf file mistakenly made it seem as though
we supported this setting to be false. Right now, Director only supports
this setting as true, so this has been clarified.