Support JDK yum repo in faster bootstrap AMI build
The build-ami.sh script now accepts the choice of a yum repository where
a JDK RPM resides; this allows the AMI build process to retrieve an RPM
from the Cloudera Director repo, which includes an RPM for Java 8. The
Director repo is the default.

The package name for the Java 8 RPM is changed to be the same as the one
available from the Cloudera Director repository.